tesseract/requirements.txt
opencv-python
pillow
google-cloud-storage
google-auth
